refactor: cleanup the dependency chain - ServerContext should be the only thing being public in server mod
